move DynamicType to mocks
