+++ deque/.gitignore
@@ -14,4 +14,3 @@ Release/
 Bin/
 
 .idea/
-
