--- deque/data/three.memcheck/answer.txt
@@ -1,39 +1,0 @@
-Deque CheckTool Package Version 1.5 Offical Version
-
----------------------------------------------------------------------------
-Test Zone A: Correctness Testing...
-Test 1: Push Series -> push_back operation testing...              PASSED
-Test 2: Push Series -> push_front operation testing...             PASSED
-Test 3: Pop Series -> pop_back operation testing...                PASSED
-Test 4: Pop Series -> pop_front operation testing...               PASSED
-Test 5: Push Series -> insert operation testing...                 PASSED
-Test 6: Pop Series -> erase operation testing...                   PASSED
-Test 7: Visitation Series -> at operation testing...               PASSED
-Test 8: Visitation Series -> constantly at operation testing...    PASSED
-Test 9: Visitation Series -> [] operator testing...                PASSED
-Test 10: Visitation Series -> constantly [] operator testing...     PASSED
-Test 11: Visitation Series -> front operation testing...            PASSED
-Test 12: Visitation Series -> back operator testing...              PASSED
-Test 13: Iterator Series -> +n operator testing...                  PASSED
-Test 14: Iterator Series -> -n operator testing...                  PASSED
-Test 15: Iterator Series -> +=n operator testing...                 PASSED
-Test 16: Iterator Series -> -=n operator testing...                 PASSED
-Test 17: Iterator Serise: Prefix ++ operator testing...             PASSED
-Test 18: Iterator Series -> Prefix -- operator testing...           PASSED
-Test 19: Iterator Series -> Suffix ++ operator testing...           PASSED
-Test 20: Iterator Series -> Suffix -- operator testing...           PASSED
-Test 21: Iterator Series -> - operator testing...                   PASSED
-Test 22: Iterator Series -> Iterator insert persistence testing...  PASSED
-Test 23: Iterator Series -> Iterator erase persistence testing...   PASSED
-Test 24: Robustness Series -> Invalid visitation by []...           PASSED
-Test 25: Robustness Series -> Invalid visitation by at...           PASSED
-Test 26: Robustness Series -> Invalid visitation by iterator...     PASSED
-Test 27: Robustness Series -> Invalid insert operation...           PASSED
-Test 28: Robustness Series -> Invalid erase operation...            PASSED
-Test 29: Basic Series -> size function testing...                   PASSED
-Test 30: Basic Series -> empty function testing...                  PASSED
-Test 31: Independence Series -> Copy constructor testing...         PASSED
-Test 32: Independence Series -> Operator = testing...               PASSED
-Test 33: Memory Series -> Memory leak testing...                    PASSED
-Test 34: Synthesis Series -> Comprehensive testing...               PASSED
----------------------------------------------------------------------------
