--- deque/data/two.memcheck/answer.txt
@@ -1,10 +1,0 @@
-Testing push...                         Passed
-Testing pop...                          Passed
-Testing insert...                       Passed
-Testing iterator...                     Passed
-Testing erase...                        Passed
-Testing copy and clear...               Passed
-Testing memory...                       Passed
-Final test without mercy...             Passed
-
-Congratulations, your deque passed all the tests!
